+++ src/main/java/com/afitnerd/tnra/repository/UserRepository.java
@@ -11,4 +11,5 @@ public interface UserRepository extends CrudRepository<User, Long> {
     User findByEmail(String email);
     Iterable<User> findByActiveTrue();
     List<User> findByActiveTrueOrderByFirstName();
+    List<User> findAllByOrderByActiveDescFirstNameAsc();
 }